tests: test the generated builders
Adorn a struct with #[non_exhaustive] to showcase and test how
the builder API (and Default) can be used to forward-compatibly
construct values of generated message types,
while the struct initializer syntax is forbidden.

Some of the doc tests showcase working around name conflicts between
the associated function generated for the builder API and a getter
method generated by the Message derive for a field named `builder`.
Also verify the no-conflict case where the field does not have a getter
generated for it, so the `builder` associated fn is fine.

The test crates in tests-2015 and tests-no-std are uniformly named
"tests" to simplify boilerplate in the doc tests.